arg-codex-gpt51/pyproject.toml
[project]
name = "arg-without-specs-codex-gpt51-v6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-openai>=0.1.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "neo4j>=5.15.0",
    "numpy>=1.26.0",
    "rapidfuzz>=3.5.0",
    "pytest>=8.2.0",
    "dashscope>=1.25.2",
    "websocket-client>=1.9.0",
    "httpx[socks]>=0.28.1",
    "langchain-experimental>=0.0.65",
    "json-repair>=0.54.1",
]

[project.scripts]
rag-cli = "main:cli"
